+++ src/irvine/oeis/a068/A068942.java
@@ -14,6 +14,6 @@ public class A068942 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A068942() {
-    super(0, n -> Integers.SINGLETON.sum(0, n * n, k -> Functions.FACTORIAL.z(k).multiply(Functions.STIRLING2.z(n * n, k))));
+    super(0, n -> Integers.SINGLETON.sum(0, n * n, k -> Functions.FACTORIAL.z(k).multiply(Functions.STIRLING2.z((long) n * n, k))));
   }
 }
