+++ src/irvine/oeis/a076/A076042.java
@@ -11,6 +11,6 @@ public class A076042 extends TupleTransformSequence {
 
   /** Construct the sequence. */
   public A076042() {
-    super(0, (n, s) -> s[0].compareTo(n * n) < 0 ? s[0].add(n * n) : s[0].subtract(n * n), "0", PREVIOUS);
+    super(0, (n, s) -> s[0].compareTo((long) n * n) < 0 ? s[0].add((long) n * n) : s[0].subtract((long) n * n), "0", PREVIOUS);
   }
 }
