+++ src/irvine/oeis/a113/A113128.java
@@ -12,6 +12,6 @@ public class A113128 extends LambdaTriangle {
 
   /** Construct the sequence. */
   public A113128() {
-    super(0, 1, 1, (n, k) -> Z.valueOf((n > 3 + k) ? 0 : (new int[] {1, 3, 3, 1})[n - k] * k));
+    super(0, 1, 1, (n, k) -> Z.valueOf((n > 3 + k) ? 0 : (long) (new int[]{1, 3, 3, 1})[n - k] * k));
   }
 }
