+++ src/irvine/oeis/a118/A118196.java
@@ -13,6 +13,6 @@ public class A118196 extends MultiTransformSequence {
 
   /** Construct the sequence. */
   public A118196() {
-    super(0, (self, n) -> Integers.SINGLETON.sum(0, n - 1, j -> Z.TWO.pow(j * (n - j)).multiply(self.a(j))).negate(), "1, -1");
+    super(0, (self, n) -> Integers.SINGLETON.sum(0, n - 1, j -> Z.TWO.pow((long) j * (n - j)).multiply(self.a(j))).negate(), "1, -1");
   }
 }
