+++ src/irvine/oeis/a122/A122193.java
@@ -14,6 +14,6 @@ public class A122193 extends LambdaTable {
 
   /** Construct the sequence. */
   public A122193() {
-    super(1, 1, n -> new Integer[]{2, n * 2}, (n, k) -> Integers.SINGLETON.sum(0, k, i -> Z.NEG_ONE.pow(k - i).multiply(Binomial.binomial(k, i).multiply(Z.valueOf(i * (i - 1) / 2).pow(n)))));
+    super(1, 1, n -> new Integer[]{2, n * 2}, (n, k) -> Integers.SINGLETON.sum(0, k, i -> Z.NEG_ONE.pow(k - i).multiply(Binomial.binomial(k, i).multiply(Z.valueOf((long) i * (i - 1) / 2).pow(n)))));
   }
 }
